pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "priorart-agent"
version = "0.2.0"
description = "Build-vs-borrow intelligence for agentic workflows - helps AI agents discover and evaluate open source packages"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Akshat Vasisht", email = "akshat2vasisht@gmail.com"}
]
keywords = ["cli", "mcp", "prior-art", "package-discovery", "open-source", "ai-agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Build Tools",
]

dependencies = [
    "fastmcp>=0.5.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "platformdirs>=4.0",
    "httpx>=0.24.0",
    "PyGithub>=2.1.0",
    "packaging>=23.0",
    "gitpython>=3.1.30",        # CVE-2022-24439 (command injection via crafted URL) fixed in 3.1.30
    # v0.2 semantic index — HF download + sigstore verify + ONNX embedding + HNSW search
    "huggingface_hub>=0.20",
    "sigstore>=3.0",
    "fastembed>=0.3",
    "usearch>=2.12",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pytest-cov>=4.0",
]
index-build = [
    "google-cloud-bigquery>=3.0",
]

[project.scripts]
priorart = "priorart.cli:main"
priorart-server = "priorart.server:main"

[project.urls]
Homepage = "https://github.com/akshatvasisht/prior-art"
Repository = "https://github.com/akshatvasisht/prior-art"
Documentation = "https://github.com/akshatvasisht/prior-art#readme"
Issues = "https://github.com/akshatvasisht/prior-art/issues"

# Setuptools configuration
[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
priorart = [
    "data/*.yaml",
    "data/*.yml",
]

# Tool configurations
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP"]
ignore = ["E501", "B008", "B905"]

[tool.coverage.run]
source = ["priorart"]
omit = ["src/priorart/server.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

